+++ src/index.ts
@@ -13,6 +13,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-export * from "./decorators";
+// !!!KEEP THIS IN THAT ORDER!!!
 export * from "./types";
 export * from "./utils";
+export * from "./functions";
+export * from "./decorators";