+++ src/utils/internal.ts
@@ -13,7 +13,59 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import type { ClassOrInstance, Collection, Constructor, Nullable } from "../types/internal";
+import type { DependencyInfoArg, DependencyInfoCollectionArg, DependencyItem, DependencyItemWithInfo } from "../types";
+import type { ClassOrInstance, Collection, Constructor, Nilable, Nullable } from "../types/internal";
+
+export function createDependsOnHelpers(
+    infoOrResolver: DependencyInfoArg,
+    dependenciesOrResolver: Nilable<DependencyInfoCollectionArg>
+) {
+    const { getDependencies } = require("../functions");
+
+    const getInfo = (item: DependencyItem) => {
+        if (typeof infoOrResolver === "function") {
+            return infoOrResolver(item);
+        }
+
+        return infoOrResolver;
+    };
+
+    const getDepsCollection: (() => Collection<DependencyItemWithInfo>) =
+        dependenciesOrResolver ?
+            () => {
+                return typeof dependenciesOrResolver === "function" ?
+                    dependenciesOrResolver() :
+                    dependenciesOrResolver!;
+            } :
+            getDependencies;
+
+    const addItem = (item: DependencyItem) => {
+        const deps = getDepsCollection();
+
+        const itemToAdd: DependencyItemWithInfo = {
+            ...item,
+
+            "info": getInfo(item)
+        };
+
+        if ("push" in deps) {
+            deps.push(itemToAdd);  // array-like
+        }
+        else if ("add" in deps) {
+            deps.add(itemToAdd);  // set-like
+        }
+        else {
+            throw new TypeError("unsupported collection type");
+        }
+    };
+
+    return {
+        addItem,
+        getDepsCollection,
+        getInfo
+    };
+}
+
 
 export function getClassName(classOrInstance: ClassOrInstance<any>): Nullable<string> {
     if (isConstructor(classOrInstance)) {